[API] trivial: move get_resource / get_module handlers to state.rs
Just realized that's how it should be organized. (See events.rs)

Closes: #580
