+++ api/src/index.rs
@@ -8,7 +8,7 @@ use crate::{
     failpoint::fail_point,
     log,
     metrics::{metrics, status_metrics},
-    transactions,
+    state, transactions,
 };
 use aptos_api_types::{Error, Response};
 
@@ -30,9 +30,7 @@ pub fn routes(context: Context) -> impl Filter<Extract = impl Reply, Error = Inf
         .or(openapi_spec())
         .or(accounts::get_account(context.clone()))
         .or(accounts::get_account_resources(context.clone()))
-        .or(accounts::get_account_resource(context.clone()))
         .or(accounts::get_account_modules(context.clone()))
-        .or(accounts::get_account_module(context.clone()))
         .or(accounts::get_account_state_blob(context.clone()))
         .or(transactions::get_transaction(context.clone()))
         .or(transactions::get_transactions(context.clone()))
@@ -42,6 +40,8 @@ pub fn routes(context: Context) -> impl Filter<Extract = impl Reply, Error = Inf
         .or(transactions::create_signing_message(context.clone()))
         .or(events::get_events_by_event_key(context.clone()))
         .or(events::get_events_by_event_handle(context.clone()))
+        .or(state::get_account_resource(context.clone()))
+        .or(state::get_account_module(context.clone()))
         .or(context.health_check_route().with(metrics("health_check")))
         .with(
             warp::cors()
